45. Load Next Level
	- You can access the next Scene by adding or incrementing your Scene Index.
	- One way you can know your count on your index is by Using SceneManager.GetCountInBuildSettings.
In this, if there is no next level, the user will go back to level 1 or the very first scene. By setting some conditions here. We can check if the nextSceneIndex is equivalent to the Scene Count. If it is equal to the Scene Count, the nextSceneIndex will go back to 0 which is the very first scene or level.